requirements.txt
absl-py==0.9.0
analytics-python==1.2.9
appdirs==1.4.4
appnope==0.1.0
argon2-cffi==20.1.0
asn1crypto==1.3.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.0
attrs==20.2.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
bioc==1.3.6
biopython==1.78
bleach==3.2.1
blis==0.4.1
bokeh==2.2.3
boto3==1.9.130
botocore==1.12.130
breadability==0.1.20
cachetools==4.1.1
catalogue==1.0.0
certifi==2019.3.9
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-log==0.3.2
cloudpickle==1.6.0
colorama==0.4.4
compare-mt==0.2.8
conda==4.8.3
conda-package-handling==1.6.0
cryptography==2.8
cycler==0.10.0
cymem==2.0.3
Cython==0.27.3
cytoolz==0.9.0.1
decorator==4.4.0
defusedxml==0.6.0
descartes==1.1.0
digicirc-service==0.0.1
dill==0.2.9
distlib==0.3.1
Django==1.10.8
docopt==0.6.2
docutils==0.15.2
dpr==0.1.0
el-core-news-lg==2.3.0
el-core-news-sm==2.3.0
en-core-sci-lg==0.2.4
en-core-web-md==2.3.1
en-core-web-sm==2.3.1
entrezpy==2.1.1
entrypoints==0.3
et-xmlfile==1.0.1
faiss-cpu==1.6.3
fasttext==0.9.2
filelock==3.0.12
Flask==1.1.1
Flask-Cors==3.0.8
future==0.18.2
gast==0.3.3
gensim==3.8.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gradio==1.0.3
greenlet==0.4.17
grpcio==1.30.0
h5py==2.10.0
hdbscan==0.8.26
hyperopt==0.1.2
idna==2.8
imageio==2.9.0
importlib-metadata==1.7.0
ipdb==0.13.2
ipykernel==5.3.4
ipython==7.6.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
jieba==0.42.1
Jinja2==2.11.2
jmespath==0.9.4
joblib==0.13.2
jsonlines==1.2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
langdetect==1.0.8
lebleu==0.0.1
llvmlite==0.34.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.2.2
mecab-python3==0.996.5
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==2.0.0
more-itertools==8.5.0
mplcursors==0.3
msgpack==0.6.2
msgpack-numpy==0.4.7.1
murmurhash==0.28.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
neo4j-driver==1.4.0
nest-asyncio==1.4.2
networkx==2.4
neuralcoref==4.0
nltk==3.5
nose==1.3.7
notebook==6.1.4
numba==0.51.0
numpy==1.19.4
oauthlib==3.1.0
OntologyAlignmentWithEmbeddings==1.0
openpyxl==3.0.3
opt-einsum==3.2.1
Owlready2==0.25
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.3
paramiko==2.7.1
parso==0.5.1
pathlib==1.0.1
pbr==5.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.2.0
plac==0.9.6
plotly==4.12.0
pluggy==0.13.1
portalocker==1.7.0
preshed==3.0.2
prettytable==0.7.2
prometheus-client==0.8.0
prompt-toolkit==2.0.9
protobuf==3.12.2
ptyprocess==0.6.0
pubmed-parser==0.2.2
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.5.0
pycosat==0.6.3
pycountry==20.7.3
pycparser==2.19
pyemd==0.5.1
Pygments==2.4.2
pymed==0.8.9
pymongo==3.11.0
PyNaCl==1.4.0
pyodbc==4.0.30
pyOpenSSL==19.1.0
pyparsing==2.4.2
pyrouge==0.1
pyrsistent==0.17.3
PySocks==1.7.1
pytest==3.10.1
pytest-cov==2.10.1
python-dateutil==2.8.0
pytools==2018.5.2
pytz==2013b0
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
rake-nltk==1.0.4
regex==2017.4.5
requests==2.21.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
ruamel-yaml==0.15.87
s3transfer==0.2.0
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
Shapely==1.7.0
six==1.12.0
smart-open==2.2.0
soupsieve==2.0.1
spacy==2.3.2
srsly==1.0.2
subword-nmt==0.3.7
sumy==0.8.1
tabulate==0.8.7
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
textsplit==0.5
thinc==7.4.1
threadpoolctl==2.1.0
tifffile==2020.7.17
tika==1.19
tokenizers==0.8.1rc2
toolz==0.11.1
torch==1.6.0
tornado==6.0.4
tqdm==4.51.0
traitlets==4.3.2
transformers==3.3.1
typing-extensions==3.7.4.3
ujson==4.0.1
umap==0.1.1
umap-learn==0.4.6
Unidecode==1.1.1
urllib3==1.24.2
virtualenv==20.0.33
wasabi==0.8.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.1
wget==3.2
widgetsnbextension==3.5.1
wmd==1.3.2
wrapt==1.12.1
xgboost==1.2.1
xlrd==1.2.0
zipp==3.1.0
